+++ client/task/task.go
@@ -2,7 +2,6 @@ package task
 
 import (
 	"sort"
-	"strconv"
 	"time"
 
 	log "github.com/sirupsen/logrus"
@@ -182,36 +181,4 @@ func (t taskRecord) IsCompleted() bool {
 
 func (t taskRecord) ID() (string, string, string) {
 	return t.RawID, t.TaskseriesID, t.ListID
-}
-
-type TaskPriority int
-
-const (
-	HighPriority TaskPriority = 1
-	MedPriority  TaskPriority = 2
-	LowPriority  TaskPriority = 3
-	NoPriority   TaskPriority = 4
-)
-
-func (p TaskPriority) String() (v string) {
-	switch p {
-	case HighPriority:
-		v = "(1)"
-	case MedPriority:
-		v = "(2)"
-	case LowPriority:
-		v = "(3)"
-	case NoPriority:
-		v = ""
-	}
-	return
-}
-
-func (t taskRecord) Priority() TaskPriority {
-	priority, err := strconv.Atoi(t.RawPriority)
-	if err != nil {
-		return NoPriority
-	}
-
-	return TaskPriority(priority)
 }
