+++ go.mod
@@ -1,6 +1,7 @@
 module github.com/dishbreak/gomilk
 
 require (
+	github.com/kyokomi/emoji v2.1.0+incompatible
 	github.com/pkg/browser v0.0.0-20180916011732-0a3d74bf9ce4
 	github.com/sirupsen/logrus v1.4.2
 	github.com/stretchr/testify v1.2.2
